requirements.txt
# Core Dependencies
SpeechRecognition>=3.10.0    # Speech-to-text functionality
pyaudio>=0.2.11              # Microphone access
fastmcp>=0.1.0               # MCP server framework

# ElevenLabs Integration (Primary)
elevenlabs>=1.0.0            # ElevenLabs API client
websockets>=11.0.0           # For real-time streaming
aiohttp>=3.8.0               # Async HTTP client

# Speech Recognition Backends (Fallback)
openai-whisper>=20230918     # Local Whisper models
torch>=2.0.0                 # For Whisper GPU acceleration (optional)

# Utilities
python-dotenv>=1.0.0         # Environment variable management
pydub>=0.25.1               # Audio processing
numpy>=1.24.0               # Audio data manipulation

# Development Tools
pytest>=7.4.0               # Testing framework
pytest-asyncio>=0.21.0      # Async test support
black>=23.0.0               # Code formatting
flake8>=6.0.0              # Linting

# Optional (for future sprints)
# openai>=1.0.0             # For OpenAI Realtime API (Sprint 7)
# websockets>=11.0          # For streaming recognition
# rich>=13.0.0              # Better CLI output